refactor(interactive-section): extract section-registry module [Pattern I]
Phase: Tier A
Step: A1
Pattern: I (module-level registry + lifecycle guards)
Risk: low

Why: The module-level step registry (counter, Map, totalDocumentSteps,
documentStepOffsets, autoDocumentOrder) plus its three public functions
(registerSectionSteps, getTotalDocumentSteps, getDocumentStepPosition)
are pure bookkeeping with order-sensitive offset math. Per the
Guidelines (Pattern I — "Extract registry/guard owner before splitting
consumers"), the registry's owner moves first. interactive-section.tsx
still owns resetInteractiveCounters() because it must call the seven
per-step-type counter resets, which would otherwise force the registry
module to import every step module.

The new module exposes:
  - resetRegistry() — clears counter + map + total + offsets +
    autoDocumentOrder.
  - nextSectionCounter() — increments + returns; used by the fallback
    section-ID generator in InteractiveSection.
  - registerSectionSteps / getTotalDocumentSteps / getDocumentStepPosition
    — semantics unchanged.

interactive-section.tsx re-exports the three public functions so
content-renderer.tsx and use-standalone-persistence.ts continue to
import from the same path. No consumer migration in this commit
(Principle 6: separate moves from migrations).

Tripwires verified:
  - section-registry.tripwire.test.ts: all 6 cases pass (offsets,
    documentOrder precedence, re-registration preserves order,
    auto-incrementing fallback, reset, unknown-section lookup).
  - All other interactive-tutorial suites pass unchanged.

Checkpoint:
  - npm run typecheck: pass
  - npx eslint: pass
  - npx prettier --check: pass
  - npx jest src/components/interactive-tutorial: 16 suites pass,
    246 tests pass + 5 todos + 12 pre-existing skips
  - interactive-section.tsx line delta: -79 / +18 (~61 lines reduced)

Notes:
  - interactiveSectionCounter is no longer module-private to
    interactive-section.tsx; nextSectionCounter() is the only way to
    advance it. This narrows the surface and makes the counter testable
    in isolation as part of the registry tripwire.
  - The Map/Set re-allocation pattern in resetRegistry mirrors the
    original .clear() semantics rather than re-assignment, preserving
    referential stability for any consumer that captured a reference.
